Fix for bug in test, adds Django dependency. Bumps version number to 0.6.
This commit adds an install dependency for Django 1.4+ - which is considered
acceptable as it's a Django app - no django, no go. I have also added a cut-
down manage.py and settings file so that the django test runner can be used
to run the tests from the repo, outside of any real django project. This
makes it much easier to develop on.

Removes Highrise settings from test_settings file.